# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/investors:
    get:
      operationId: listAdminInvestors
      tags: [admin]
      summary: List all investors (admin only)
      responses:
        "200":
          description: List of investors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvestorAdmin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/investors/{id}/status:
    patch:
      operationId: updateInvestorStatus
      tags: [admin]
      summary: Update investor status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestorStatusInput"
      responses:
        "200":
          description: Updated investor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorAdmin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    InvestorAdmin:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - status
        - createdAt
...
    InvestorStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
      required:
        - status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error