# lib/api-zod/src/generated/types/investorAdmin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestorAdminStatus } from './investorAdminStatus';

export interface InvestorAdmin {
  id: number;
  fullName: string;
  email: string;
  status: InvestorAdminStatus;
  createdAt: Date;
}