# lib/api-zod/src/generated/types/investorStatusInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvestorStatusInputStatus } from './investorStatusInputStatus';

export interface InvestorStatusInput {
  status: InvestorStatusInputStatus;
}